refactor: Restructuring the project
