+++ src/routes/api/getColumnMapping.js
@@ -1,0 +1,11 @@
+'use strict'
+
+module.exports = (app) => {
+  app.get('/getColumnMapping', (_req, res) => {
+    // if (Object.keys(clientColumnMapping).length > 0) {
+    //   res.send(JSON.stringify(clientColumnMapping));
+    // } else {
+    res.send('Column Mapping not set yet ...');
+    // }
+  });
+}
